Begin update quantum/_common_ and quantum/quantumOps.py
Updates the Bra class in quantumOps.py and begins updating old
quantum/common.py to the quantum/_common_.ipynb format. Creates a backup
copy of the old quantum/quantumOps.py imaginatively named
quantum/quantumOps_old.py. Related Issues: #97 .
